Fixed migraation issues -alembic couldnt locate public schema and rename tables -good conventions

+++ server/models/user.py
@@ -9,7 +9,7 @@ class User(db.Model):
     last_name = Column(String(120), nullable=False)
     email = Column(String(255), unique=True, nullable=False)
     phone = Column(String(50), unique=True, nullable=False)
-    id_number=Column(Integer(15),unique=True,nullable=False)
+    id_number=Column(Integer,unique=True,nullable=False)
     password = Column(String(255), nullable=False)
     active=Column(Boolean,nullable=False,default=True)
